Address review feedback
Toolchain:
- Keep file_prefix_map enabled for Android (hermetic working-dir remap) instead
  of disabling it; the worker only requires DEVELOPER_DIR for the Apple
  developer-dir remap, so guard that remap on DEVELOPER_DIR being set (a cross
  compile on a macOS host has none). Fixes non-hermetic debug info.
- Extract the Android link logic into _swift_android_linkopts_cc_info keyed on
  os == "android"; drop the redundant -lm (the clang driver links libm anyway;
  -lstdc++ stays, since it selects the shared libc++_shared.so the SDK intends).
- Rename the SDK repo's host_swiftc attr to paired_swiftc.
- Trim verbose comments/docs across the toolchain and extensions; simplify the
  clang-only error message.

Example (//examples/cross_compilation/android_app):
- Consolidate the Swift sources, Kotlin app, and BUILD into android_app/.
- Build the .so as a dep of android_binary (its native split configures it for
  Android) instead of a hand-written platform transition; use the platforms
  rules_android exposes (@rules_android//:arm64-v8a) rather than a custom one.
- Drop the manual tags; mark the Android-only targets target_compatible_with
  os:android so the //examples/... wildcard skips them on a host and the example
  runs in the normal macOS CI job. Remove the dedicated cross-compilation task.
- Split the example's dev dependencies into android_app/android.MODULE.bazel,
  include()d from the root MODULE; move the Maven lock out of the repo root;
  link rules_android#485 for the maven.install requirement.
- Set the rules_android flags (android_platforms + hermetic JDK) globally rather
  than behind a --config.
