--- examples/cross_compilation/BUILD.bazel
@@ -1,45 +1,0 @@
-load("//examples/embedded:transition.bzl", "transition_binary")
-load("//swift:swift.bzl", "swift_binary", "swift_library")
-
-package(default_visibility = ["//visibility:public"])
-
-# The platform covered by the Android Swift SDK toolchain registered by this
-# repository's dev `MODULE.bazel` (via the `swift` extension's `android_sdk`
-# tag). Build the targets below with `--platforms` set to this.
-platform(
-    name = "android-aarch64",
-    constraint_values = [
-        "@platforms//cpu:aarch64",
-        "@platforms//os:android",
-    ],
-)
-
-# A plain library dependency, compiled for whichever platform depends on it,
-# demonstrating that a normal `swift_library` is reused unchanged across targets.
-swift_library(
-    name = "Greeter",
-    srcs = ["Sources/Greeter/Greeter.swift"],
-    module_name = "Greeter",
-    tags = ["manual"],
-)
-
-# ---------------------------------------------------------------------------
-# Android: a JNI shared library, loaded by Kotlin via `System.loadLibrary`.
-# ---------------------------------------------------------------------------
-
-# `linkshared` produces `libSwiftJNI.so`. The JNI entry point is written in
-# Swift (it `import`s the SDK's `Android` module for the JNI types) and calls
-# into the `Greeter` library. See `android_app/` for the APK that loads it.
-swift_binary(
-    name = "SwiftJNI",
-    srcs = ["Sources/SwiftJNI/SwiftJNI.swift"],
-    linkshared = True,
-    tags = ["manual"],
-    deps = [":Greeter"],
-)
-
-transition_binary(
-    name = "libSwiftJNI.so",
-    binary = ":SwiftJNI",
-    platform = ":android-aarch64",
-)
