+++ examples/cross_compilation/android_app/Sources/Greeter/Greeter.swift
@@ -1,0 +1,11 @@
+public struct Greeter {
+  private let subject: String
+
+  public init(subject: String) {
+    self.subject = subject
+  }
+
+  public func greeting() -> String {
+    return "Hello from Swift, \(subject)!"
+  }
+}
