+++ examples/cross_compilation/android_app/android.MODULE.bazel
@@ -1,0 +1,68 @@
+# Dev dependencies for the runnable APK example (//examples/cross_compilation/
+# android_app), included from the root MODULE.bazel. rules_android + rules_kotlin
+# package the APK the real-world way; they pull in a Maven graph and are heavy,
+# so they're kept out of the root module and are all dev_dependency. Consumers of
+# rules_swift never see them; a consumer brings their own Android rules.
+#
+# `android_ndk` (the hermetic_android_toolchains extension) and `@androidndk` are
+# defined in the root module above this include().
+
+bazel_dep(name = "rules_android", version = "0.7.3", dev_dependency = True)
+bazel_dep(name = "rules_java", version = "9.3.0", dev_dependency = True)
+bazel_dep(name = "rules_kotlin", version = "2.2.2", dev_dependency = True)
+bazel_dep(name = "rules_jvm_external", version = "7.0", dev_dependency = True)
+
+# rules_java's hermetic JDK + java tools for rules_android's Java actions.
+rules_java_toolchains = use_extension(
+    "@rules_java//java:extensions.bzl",
+    "toolchains",
+    dev_dependency = True,
+)
+use_repo(rules_java_toolchains, "remote_java_tools")
+
+# rules_android's own Maven dependencies, pinned to a checked-in lock file:
+#   REPIN=1 bazel run @rules_android_maven//:pin
+# rules_android requiring a consumer-side maven.install is a known wart; see
+# https://github.com/bazelbuild/rules_android/issues/485.
+maven = use_extension(
+    "@rules_jvm_external//:extensions.bzl",
+    "maven",
+    dev_dependency = True,
+)
+maven.install(
+    name = "rules_android_maven",
+    aar_import_bzl_label = "@rules_android//rules:rules.bzl",
+    artifacts = [],
+    known_contributing_modules = ["rules_android"],
+    lock_file = "//examples/cross_compilation/android_app:rules_android_maven_install.json",
+    repositories = [
+        "https://maven.google.com",
+        "https://repo1.maven.org/maven2",
+    ],
+    use_starlark_android_rules = True,
+)
+use_repo(maven, "rules_android_maven")
+
+# Point rules_android at the hermetic @androidsdk created by the android_ndk.sdk
+# tag in the root module (instead of requiring a local $ANDROID_HOME). The
+# extension proxy is re-declared here because include() segments don't share
+# names with the root module; this resolves to the same extension usage.
+android_ndk = use_extension(
+    "@hermetic_android_toolchains//:extensions.bzl",
+    "android",
+    dev_dependency = True,
+)
+use_repo(android_ndk, "androidsdk")
+
+rules_android_sdk = use_extension(
+    "@rules_android//rules/android_sdk_repository:rule.bzl",
+    "android_sdk_repository_extension",
+    dev_dependency = True,
+)
+
+override_repo(rules_android_sdk, "androidsdk")
+
+register_toolchains(
+    "@androidsdk//:all",
+    dev_dependency = True,
+)
