+++ examples/cross_compilation/android_app/java/com/example/swiftjni/NativeBridge.kt
@@ -1,9 +1,8 @@
 package com.example.swiftjni
 
 /**
- * Loads the Swift JNI shared library (`libSwiftJNI.so`, built by the
- * `//examples/cross_compilation:SwiftJNI` `swift_binary(linkshared = True)`)
- * and exposes its Swift entry point to Kotlin.
+ * Loads the Swift JNI shared library (`libSwiftJNI.so`, built by the `:SwiftJNI`
+ * `swift_binary(linkshared = True)`) and exposes its Swift entry point to Kotlin.
  *
  * The native method binds by name to the Swift `@_cdecl` function
  * `Java_com_example_swiftjni_NativeBridge_greetingFromSwift`, which in turn
