+++ examples/cross_compilation/android_app/rules_android_maven_install.json
