+++ tools/worker/swift_runner.cc
@@ -587,10 +587,18 @@ bool SwiftRunner::ProcessArgument(
     consumer(std::filesystem::current_path().string() + "=" + new_path);
 
 #if __APPLE__
-    std::string developer_dir = "__BAZEL_XCODE_DEVELOPER_DIR__";
-    if (bazel_placeholder_substitutions_.Apply(developer_dir)) {
-      consumer(flag);
-      consumer(developer_dir + "=/PLACEHOLDER_DEVELOPER_DIR");
+    // The developer-directory remap only applies to toolchains that use the
+    // Apple developer directory. A cross-compile (e.g. to Android) running on a
+    // macOS host has no DEVELOPER_DIR; skip the remap rather than hard-failing,
+    // keeping the working-directory remap above so file_prefix_map stays
+    // hermetic for those targets.
+    const char* developer_dir_env = std::getenv("DEVELOPER_DIR");
+    if (developer_dir_env != nullptr && developer_dir_env[0] != '\0') {
+      std::string developer_dir = "__BAZEL_XCODE_DEVELOPER_DIR__";
+      if (bazel_placeholder_substitutions_.Apply(developer_dir)) {
+        consumer(flag);
+        consumer(developer_dir + "=/PLACEHOLDER_DEVELOPER_DIR");
+      }
     }
 #endif
   };
